requirements.txt
streamlit
requests
pandas
openai
python-dotenv
matplotlib